rename db_functions to db_queries
